userDefineRole & checkAnswer
defineUserRole:
switch statement handled to allow user to proceed to appropriate menu or get kicked out if they're not a manager or provider

checkAnswer:
(in the utility class)
Converted return type to int to allow for a more robust error check. It now handles all types of input: yes, no, or invalid input.