Changed changeMemberStanding() and promptInt()
Got rid of suspendMember() and unsuspendMember() functions in Manager
and added a call to appropriate DataAccess suspend/unsuspend call in
changeMemberStanding(). Having those other two functions in Manager was
unneccessary for the reason given in prior sentence. Fixed promptInt()
to perform Integer.parseInt() inside the function. Removed that code
where we were prompting for the zip code and added call to promptInt()
to get that information. Shortened lines of code used in Manager
significantly.